ift_tool/requirements.txt
openai
pydantic
pyyaml
python-dotenv